build login register logic is done
build user context because this app wont be huge and we do not need complex solutions like redux.
we handle forms with useCallback to momo the data end boost performance
user is added to localstorage if successfully log in
user is successfully removed if logout appear in the events
add error handling on both login and register.